CM - Ajustes pós-merge para adequação da estrutura.
