+++ src/SME.SGP.Aplicacao/CasosDeUso/Ocorrencia/ExcluirOcorrenciaUseCase.cs
@@ -1,5 +1,7 @@
 ﻿using MediatR;
+using SME.SGP.Infra;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace SME.SGP.Aplicacao
@@ -10,9 +12,22 @@ namespace SME.SGP.Aplicacao
         {
         }
 
-        public async Task<bool> Executar(IEnumerable<long> lstIds)
+        public async Task<bool> Executar(IEnumerable<long> ids)
         {
-            var retorno = await mediator.Send(new ExcluirOcorrenciaCommand(lstIds));
+            var retorno = new RetornoBaseDto();
+            if (!ids?.Any() ?? true)
+            {
+                retorno.Mensagens.Add("Devem ser informadas ao menos uma ocorrência para exclusão.");
+                return retorno;
+            }
+
+            foreach (var id in ids)
+            {
+                var retornoExclusao = await mediator.Send(new ExcluirOcorrenciaCommand { Id = id });
+                if (retornoExclusao.ExistemErros)
+                    retorno.Mensagens.AddRange(retornoExclusao.Mensagens);
+            }
+
             return retorno;
         }
     }
