--- src/SME.SGP.Aplicacao/Interfaces/CasosDeUso/Ocorrencias/IExcluirOcorrenciasPorCodigoUseCase.cs
@@ -1,9 +1,0 @@
-﻿using SME.SGP.Infra;
-using System.Collections.Generic;
-
-namespace SME.SGP.Aplicacao
-{
-    public interface IExcluirOcorrenciasPorCodigoUseCase : IUseCase<IEnumerable<long>, RetornoBaseDto>
-    {
-    }
-}
